course-mapper/course_mapper/etl/satellite_processing.py
"""
Satellite imagery processing and course feature segmentation.

This module handles:
- Downloading satellite imagery tiles for course areas
- Segmenting course features (greens, fairways, bunkers, etc.)
- Converting segmentation masks to polygons for storage
"""
import logging
from typing import Dict, List, Tuple, Optional
import numpy as np
from shapely.geometry import Polygon, MultiPolygon
import cv2

from course_mapper.config import settings

logger = logging.getLogger(__name__)


class SatelliteProcessor:
    """Process satellite imagery to extract golf course features."""
    
    def __init__(self, api_key: Optional[str] = None):
        self.api_key = api_key or settings.satellite_api_key
    
    def download_imagery_tiles(
        self,
        bounds: Tuple[float, float, float, float],  # (min_lat, min_lon, max_lat, max_lon)
        zoom_level: int = 18,
        provider: str = "mapbox"  # or "google", "arcgis"
    ) -> Optional[np.ndarray]:
        """
        Download satellite imagery tiles for a bounding box.
        
        Args:
            bounds: Bounding box (min_lat, min_lon, max_lat, max_lon)
            zoom_level: Tile zoom level (higher = more detail)
            provider: Imagery provider name
            
        Returns:
            Combined image array, or None if download failed
            
        Note:
            This is a stub - implement actual tile fetching based on provider API.
        """
        min_lat, min_lon, max_lat, max_lon = bounds
        
        logger.info(f"Downloading imagery for bounds: {bounds}")
        logger.warning("⚠️  Stub implementation - replace with real provider API")
        
        # TODO: Implement tile fetching
        # Example for Mapbox:
        # - Convert lat/lon to tile coordinates
        # - Fetch tiles via Mapbox Static API
        # - Stitch tiles together
        # - Return as numpy array
        
        # Stub: Return a placeholder
        return None
    
    def segment_course_features(self, image: np.ndarray) -> Dict[str, np.ndarray]:
        """
        Segment golf course features from satellite imagery.
        
        Uses computer vision/ML to identify:
        - Greens (putting surfaces)
        - Fairways
        - Bunkers
        - Water hazards
        - Rough areas
        
        Args:
            image: RGB image array (H, W, 3)
            
        Returns:
            Dictionary mapping feature type to binary mask array
        """
        logger.info("Segmenting course features from imagery")
        logger.warning("⚠️  Stub implementation - replace with real ML model")
        
        # TODO: Implement segmentation
        # Options:
        # 1. Pre-trained semantic segmentation model (TensorFlow/PyTorch)
        # 2. Rule-based CV with color thresholding + morphological ops
        # 3. Hybrid: ML for greens/fairways, CV for water/bunkers
        
        # Stub: Return empty masks
        height, width = image.shape[:2]
        
        return {
            'green': np.zeros((height, width), dtype=np.uint8),
            'fairway': np.zeros((height, width), dtype=np.uint8),
            'bunker': np.zeros((height, width), dtype=np.uint8),
            'water': np.zeros((height, width), dtype=np.uint8),
            'rough': np.zeros((height, width), dtype=np.uint8),
            'tee_box': np.zeros((height, width), dtype=np.uint8)
        }
    
    def mask_to_polygons(
        self,
        mask: np.ndarray,
        transform: Optional[Tuple] = None,
        min_area: float = 10.0
    ) -> List[Polygon]:
        """
        Convert a binary mask to shapely polygons.
        
        Args:
            mask: Binary mask array (H, W)
            transform: Optional geotransform tuple (from rasterio)
            min_area: Minimum polygon area in pixels to include
            
        Returns:
            List of Polygon objects in image coordinates
        """
        # Find contours
        contours, _ = cv2.findContours(
            mask.astype(np.uint8),
            cv2.RETR_EXTERNAL,
            cv2.CHAIN_APPROX_SIMPLE
        )
        
        polygons = []
        for contour in contours:
            if len(contour) < 3:
                continue
            
            # Calculate area
            area = cv2.contourArea(contour)
            if area < min_area:
                continue
            
            # Convert contour to polygon
            points = contour.reshape(-1, 2)
            polygon = Polygon(points)
            
            if polygon.is_valid and polygon.area > min_area:
                polygons.append(polygon)
        
        return polygons
    
    def georeference_polygons(
        self,
        polygons: List[Polygon],
        bounds: Tuple[float, float, float, float],
        image_shape: Tuple[int, int]
    ) -> List[Polygon]:
        """
        Convert image-coordinate polygons to geographic coordinates.
        
        Args:
            polygons: List of polygons in image coordinates
            bounds: Geographic bounds (min_lat, min_lon, max_lat, max_lon)
            image_shape: (height, width) of the source image
            
        Returns:
            List of polygons in geographic coordinates (lat, lon)
        """
        min_lat, min_lon, max_lat, max_lon = bounds
        height, width = image_shape
        
        lat_range = max_lat - min_lat
        lon_range = max_lon - min_lon
        
        geopolygons = []
        for poly in polygons:
            # Transform coordinates
            coords = []
            for x, y in poly.exterior.coords:
                # Image coordinates to geographic
                lon = min_lon + (x / width) * lon_range
                lat = max_lat - (y / height) * lat_range  # Y is flipped
                coords.append((lon, lat))
            
            geopoly = Polygon(coords)
            if geopoly.is_valid:
                geopolygons.append(geopoly)
        
        return geopolygons
    
    def process_course_imagery(
        self,
        bounds: Tuple[float, float, float, float],
        course_id: str
    ) -> Dict[str, List[Polygon]]:
        """
        Complete pipeline: download imagery, segment, convert to polygons.
        
        Args:
            bounds: Geographic bounds
            course_id: UUID of the course
            
        Returns:
            Dictionary mapping feature type to list of polygons
        """
        # Download imagery
        image = self.download_imagery_tiles(bounds)
        if image is None:
            logger.warning(f"Could not download imagery for course {course_id}")
            return {}
        
        # Segment features
        masks = self.segment_course_features(image)
        
        # Convert to polygons
        result = {}
        image_shape = image.shape[:2]
        
        for feature_type, mask in masks.items():
            polygons = self.mask_to_polygons(mask)
            geopolygons = self.georeference_polygons(polygons, bounds, image_shape)
            
            if geopolygons:
                result[feature_type] = geopolygons
                logger.info(f"Extracted {len(geopolygons)} {feature_type} polygons")
        
        return result


def process_course_example():
    """
    Example usage: Process imagery for a course.
    """
    processor = SatelliteProcessor()
    
    # Example bounds (Pebble Beach hole 1)
    bounds = (36.568, -121.950, 36.571, -121.946)
    
    features = processor.process_course_imagery(bounds, "example-course-id")
    print(f"Extracted features: {list(features.keys())}")


if __name__ == "__main__":
    process_course_example()



