course-mapper/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Geospatial
geopandas==0.14.1
shapely==2.0.2
rasterio==1.3.9
fiona==1.9.5
pyproj==3.6.1

# OSM data
requests==2.31.0
overpy==0.6

# Image processing
opencv-python==4.8.1.78
numpy==1.26.2
Pillow==10.1.0

# Optional ML stubs
# tensorflow==2.15.0  # Uncomment if using TensorFlow
# torch==2.1.1  # Uncomment if using PyTorch

# Utilities
python-dotenv==1.0.0



